// planning-idsdp-pps-frontend/src/app/main/modules/dpp-tapp-management/models/DppAmortizationSchedule.model.ts
export class DppAmortizationSchedule{
    id: any;
    uuid: any;
    projectConceptUuid: String;
    projectConceptMasterId: number;
    projectName: string;
    totalInvestment: number;
    loanPortion: number;
    loanPeriod: number;
    rateOfInterest: number;
    gracePeriod: number;
    gracePeriods: any;
    loanPeriods: any;

}